@@ -1,8 +1,8 @@
 module github.com/samborkent/safe
 
-go 1.23.2
+go 1.24.0
 
 require (
-	github.com/samborkent/check v0.0.0-20241007113630-271a335cc238
-	golang.org/x/exp v0.0.0-20241004190924-225e2abe05e6
+	github.com/samborkent/check v0.0.0-20250215121809-35c000968ba0
+	golang.org/x/exp v0.0.0-20250210185358-939b2ce775ac
 )
