@@ -8,16 +8,15 @@ import (
 // TODO: test
 
 type Map[Key comparable, Value any] struct {
-	initialized bool
+	lock        sync.RWMutex
 	data        map[Key]Value
-	lock        *sync.RWMutex
+	initialized bool
 }
 
 func NewMap[Key comparable, Value any](size int) *Map[Key, Value] {
 	return &Map[Key, Value]{
 		initialized: true,
 		data:        make(map[Key]Value, size),
-		lock:        new(sync.RWMutex),
 	}
 }
 
@@ -27,8 +26,9 @@ func (m *Map[Key, Value]) Clear(key Key) {
 	}
 
 	m.lock.Lock()
+	defer m.lock.Unlock()
+
 	m.data = make(map[Key]Value, len(m.data))
-	m.lock.Unlock()
 }
 
 func (m *Map[Key, Value]) Delete(key Key) {
@@ -37,8 +37,9 @@ func (m *Map[Key, Value]) Delete(key Key) {
 	}
 
 	m.lock.Lock()
+	defer m.lock.Unlock()
+
 	delete(m.data, key)
-	m.lock.Unlock()
 }
 
 func (m *Map[Key, Value]) Len() int {
@@ -46,8 +47,6 @@ func (m *Map[Key, Value]) Len() int {
 		return 0
 	}
 
-	m.lock.RLock()
-	defer m.lock.RUnlock()
 	return len(m.data)
 }
 
@@ -57,20 +56,22 @@ func (m *Map[Key, Value]) Load(key Key) (value Value, ok bool) {
 	}
 
 	m.lock.RLock()
+	defer m.lock.RUnlock()
+
 	value, ok = m.data[key]
-	m.lock.RUnlock()
 	return value, ok
 }
 
 // TODO: test
 func (m *Map[Key, Value]) Range() iter.Seq2[Key, Value] {
 	if !m.initialized {
-		return func(func(Key, Value) bool) {
-			return
-		}
+		return func(func(Key, Value) bool) {}
 	}
 
 	return func(yield func(Key, Value) bool) {
+		m.lock.RLock()
+		defer m.lock.RUnlock()
+
 		for k, v := range m.data {
 			if !yield(k, v) {
 				return
@@ -85,9 +86,11 @@ func (m *Map[Key, Value]) Swap(key Key, value Value) (previous Value, loaded boo
 	}
 
 	m.lock.Lock()
+	defer m.lock.Unlock()
+
 	previous, loaded = m.data[key]
 	m.data[key] = value
-	m.lock.Unlock()
+
 	return previous, loaded
 }
 
@@ -97,6 +100,7 @@ func (m *Map[Key, Value]) Store(key Key, value Value) {
 	}
 
 	m.lock.Lock()
+	defer m.lock.Unlock()
+
 	m.data[key] = value
-	m.lock.Unlock()
 }
