@@ -1,7 +1,7 @@
 package safe
 
-// Dereference with nil pointer dereference protection.
-// Returns the zero value of the underlying type is the provided pointer is nil.
+// Dereference can be used to safely dereference a pointer with nil pointer dereference protection.
+// Returns the zero value of the underlying type if the provided pointer is nil.
 func Dereference[T any](pointer *T) T {
 	if pointer == nil {
 		return *new(T)
@@ -10,7 +10,8 @@ func Dereference[T any](pointer *T) T {
 	}
 }
 
-// Set the underlying value of a pointer with nil pointer dereference protection.
+// SetValue sets the underlying value of a pointer with nil pointer dereference protection.
+// No-op if the underlying pointer is nil.
 func SetValue[T any](pointer **T, value T) {
 	if pointer != nil {
 		if *pointer == nil {
